refac: separate validation from db schema
