+++ src/models/User.ts
@@ -1,30 +1,5 @@
-import { z } from 'zod';
 import { Schema, model } from 'mongoose';
-
-const oauthSchema = z.object({
-  facebook: z.string(),
-  emailFacebook: z.string().email(),
-  google: z.string(),
-  emailGoogle: z.string().email(),
-  email: z.string().email(),
-});
-
-const devicesSchema = z.object({
-  phone: z.string(),
-  token: z.string(),
-  device_id: z.string(),
-});
-
-const _userSchema = z.object({
-  ra: z.number().int(),
-  created_at: z.date(),
-  email: z.string().email(),
-  confirmed: z.boolean(),
-  devices: z.array(devicesSchema),
-  oauth: oauthSchema,
-});
-
-type IUser = z.infer<typeof _userSchema>;
+import { IUser } from '@/modules/user-schema';
 
 const userSchema = new Schema<IUser>({
   ra: {
